requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
django-environ
drf-yasg
Faker
django-cors-headers
django-filter
